+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4054,7 +4054,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KY4GW3T3VNDZPEBHMRVQNSG6"),
+  id: z.string().default("01KY4MCV537RZF1NHCXJ15XWWR"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),